+++ src/pages/commits/ActiveCommits/utils.ts
@@ -1,5 +1,5 @@
-import { mapTaskStatusToColor, sortedStatusColor } from "constants/task";
 import { MainlineCommitsQuery } from "gql/generated/types";
+import { groupStatusesByColor } from "utils/statuses";
 
 export type ColorCount = { count: number; statuses: string[]; color: string };
 
@@ -7,43 +7,6 @@ export type GroupedResult = {
   stats: ColorCount[];
   max: number;
   total: number;
-};
-
-export const groupStatusesByColor = (
-  statusCounts: { status: string; count: number }[]
-) => {
-  const counts: { [key: string]: ColorCount } = {};
-
-  statusCounts.forEach((stat) => {
-    const statusColor = mapTaskStatusToColor[stat.status];
-    if (counts[statusColor]) {
-      counts[statusColor].count += stat.count;
-      if (!counts[statusColor].statuses.includes(stat.status)) {
-        counts[statusColor].statuses.push(stat.status);
-      }
-    } else {
-      counts[statusColor] = {
-        count: stat.count,
-        statuses: [stat.status],
-        color: statusColor,
-      };
-    }
-  });
-
-  let max = -1;
-  let total = 0;
-  const stats: ColorCount[] = [];
-
-  // Populate STATS array with ColorCount object in sorted color order
-  sortedStatusColor.forEach((color) => {
-    if (counts[color]) {
-      total += counts[color].count;
-      max = Math.max(max, counts[color].count);
-      stats.push(counts[color]);
-    }
-  });
-
-  return { stats, max, total };
 };
 
 export const findMaxGroupedTaskStats = (groupedTaskStats: {
