+++ src/utils/statuses/index.ts
@@ -1,5 +1,11 @@
 import { getStatusBadgeCopy } from "./getStatusBadgeCopy";
+import { groupStatusesByColor } from "./groupStatusesByColor";
 import { isFailedTaskStatus } from "./isFailedTaskStatus";
 import { sortTasks } from "./sort";
 
-export { isFailedTaskStatus, getStatusBadgeCopy, sortTasks };
+export {
+  isFailedTaskStatus,
+  getStatusBadgeCopy,
+  sortTasks,
+  groupStatusesByColor,
+};
